create a new file networks.go and move network part codes from cluster.go into networks.go
